refactor

+++ app/controllers/queue_items_controller.rb
@@ -5,14 +5,16 @@ class QueueItemsController < ApplicationController
     @queue_items = current_user.queue_items
   end
 
-  def remove_queue_item
+  def destroy
     queue_item = QueueItem.find(params[:id])
-    QueueItem.where(
-      "position > ? AND user_id = ?", 
-      queue_item.position, current_user.id).each do |item|
-      item.decrement!(:position)
+    if current_user.queue_items.include? queue_item
+      QueueItem.where(
+        "position > ? AND user_id = ?", 
+        queue_item.position, current_user.id).each do |item|
+        item.decrement!(:position)
+      end
+      queue_item.destroy
     end
-    queue_item.destroy
     redirect_to my_queue_path
   end
 end
